requirements.txt
Pillow
django-widget-tweaks
social-auth-app-django